Set additional mongoose config in user model
These must be set here and be set on each model to avoid deprecation warnings. If passed in as options to mongoose.connect, the deprecation warnings will not be supressed and will be logged to console. Not sure why, but doing it this way is what was suggested in mongoose #6890