Archive and Unarchive button now update component automatically
I decided to combine LOADING_TRIPS & LOADING_ARCHIVED_TRIPS, I think having both added unnecessary complexity to the redux action and reducers. Both are pulling from the same resource and then filtering it before saving. This is extra noise that is better delegated to the react component that displays the data. For example, all we are doing now is a simple if(trip.isArchived) check when mapping trip to TripCard. This one line of code replaces 3 functions and 3 action types.