Change jwt.decode to jwt.verify
In the documentation for the "jsonwebtoken" library, it gives the following warning for the function jwt.decode:

https://www.npmjs.com/package/jsonwebtoken

******
Warning: This will not verify whether the signature is valid. You should not use this for untrusted messages. You most likely want to use jwt.verify instead.
******

Therefore, to ensure that the signature is valid, the jwt.verify should be called, not jwt.decode.